[FEATURE] implement ssh
New commands:

* `hark ssh`: ssh interactively into a machine
* `hark vm mappings`: show all port mappings

New database table: `port_mappings`.

The `hark vm new` command now sets up an SSH port mapping through the
driver after it creates a new machine, and it records it in the DB
through the DAL.

The keypair has been committed and added to MANIFEST.in, so that it will
be included in any distribution. Technically only the private keys is
needed, but it's useful for documentation purposes to have both here.