[FEATURE] support python 2.x
This is much more useful for the life and viability of the project than
the sugar from python 3; it should work out of the box on any reasonably
modern system, it can be distributed through pypi as a universal wheel,
etc.

There could still be some bugs in here from unicode literals; hopefully
they'll be uncovered through manual testing with python2.7 or improving
test coverage in the virtualbox driver.

This involved:

* removing typing and asyncio dependencies
* importing `absolute_import` where needed (in cli modules)
* removing annotations everywhere
* Command taking an array, not `*args`
* handle unittest.mock not being available in 2.x

Also documented supported versions in the README, and updated both the
tox and travis config for tests.