[REFACTOR] Move complex logic into new procedure package
The design of this is not yet clean, but the idea is to encapsulate a
complex set of operations on clients, drivers, contexts, etc. into a
'procedure', to get this logic out of the CLI layer. This makes it more
testable, and means that it can be reused through a different entry
point such as a HTTP server.