[FEATURE] assign static IPs to machines on the host-only interface
This is only relevant to virtualbox for now. Other drivers when added
will need a similar mechanism.

For this to work, there also needs to be a static IP created for the
host machine. This will probably break if we reuse a host-only interface
that was already there, so this branch may need a commit added whereby
hark can remember its exclusive host-only interface.

The Procedure API is refactored to be simpler.

The DAL is fixed to handle nulls. Lots of flake8 fixes.