[FIX] virtualbox: use a dedicated host-only interface
This adds a new config table which can persist global configuration like
the name of the host-only interface. With this change, hark never gets
the list of host-only interfaces from the system: it just creates one on
the first run for a context and records it in the DB through the DAL.
This means that the driver's create() method needs to be given a DAL
instance.

Resolves #8