[FEATURE] Support destroying VMs
This commit adds a new CLI command:

	hark vm destroy

This will:

* delete the VM from the driver
* delete the VM from the machine table

The driver needed support added to wait on a machine for a certain
state, so that the procedure to destroy can call stop() then wait for
the machine to be stopped before destroying it.

This commit needs to be amended to include deleting port mappings and
network interfaces.

Resolves #3